requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
pyttsx3==2.90
Requests==2.31.0
